% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRTH.R
\name{LRTH}
\alias{LRTH}
\title{The function of likelihood ratio test for genetic heteroneity}
\usage{
LRTH(x, y)
}
\arguments{
\item{x:}{a vector of length n (number of observations) for the genotypic score of a SNP; the value of gentoype score is the number of minor alleles of the SNP}

\item{y:}{a vector of length n for disease indicator; = 0 for controls, = 1 for cases.  
}
\value{
The asymptotic p-value of LRTH test
}
\description{
Take in genotypic score of a SNP , a vector of disease status and return the p-value
}

