
\name{LRTH-package}
\alias{LRTH-package}
\alias{LRTH}
\docType{package}
\title{
\packageTitle{LRTH}
}
\description{
\packageDescription{LRTH}
}

\author{
\packageAuthor{LRTH}

Maintainer: \packageMaintainer{LRTH}
}
\references{
Qian M., Shao Y., 2013. A Likelihood Ratio Test for Genome-Wide Association under Genetic Heterogeneity. Annals of Human Genetics, 77(2): 174-182.

Zhou H., Pan W., 2009. Binomial Mixture Model-based Association Tests under Genetic Heterogeneity. Annals of Human Genetics, 73(6): 614-630.}


\keyword{ Likelihood ratio test, genetic heterogeneity }