% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRTH.R
\name{LRTH}
\alias{LRTH}
\title{The function to for the likelihood ratio test for genetic heteroneity}
\usage{
LRTH(x, y)
}
\arguments{
\item{x:}{a n x 1 vector of genotypic score (i.e. 0, 1 or 2)}

\item{y:}{a n x 1 vector of disease status; case/xcontrol: 1/0}
}
\value{
The asymptotic p-value of LRTH test
}
\description{
Take in genotypic score and disease status and return the p-value
}

